Move shared Falcon test helpers
